python/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
psycopg2-binary>=2.9.6
redis>=4.5.0
django-redis>=5.2.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
dj-database-url>=2.0.0